test: rewrite tests to use ESM syntax
